takes care of point 3
Only one allowed host is enough in the container environment. All traffic is coming from Nginx.
Do not set the Host header in Nginx it will screw with the ALLOWED_HOSTS in Python.

The TUBESYNC_HOSTS now only affects the CSRF_TRUSTED_ORIGINS.
If there is a *need* the hosts are required in ALLOWED_HOSTS I would like to know in which scenario.