Log to `/config/log/nginx`
The error log contains informational and higher levels on the theory that when something goes wrong we will want as much information as we can get about why an error occurred.

Access log is compressed, in memory, then written to disk every minute. Most people will take about that long to start reading the file anyway.